+++ src/templates/Home/TaskList.tsx
@@ -1,12 +1,12 @@
 import { Box, Flex, Heading, useDisclosure } from "@chakra-ui/react";
 import { useMemo, useState } from "react";
-import { AddColumnModal } from "../AddColumnModal/AddColumnModal";
+import { AddColumnModal } from "./AddColumnModal/AddColumnModal";
 import { Column } from "./Column";
 import { Header } from "./Header";
 import { Column as ColumnType, useZustand } from "./state";
-import { AddTodoModal } from "../AddTodoModal/AddTodoModal";
+import { AddTodoModal } from "./AddTodoModal/AddTodoModal";
 import { v4 as uuid } from "uuid";
-import { addTodoAction } from "../actions/addTodoAction";
+import { addTodoAction } from "./serverActions";
 
 export function TaskList() {
   const addColumn = useZustand((store) => store.addColumn);
