--- src/templates/actions/addTodoAction.ts
@@ -1,8 +1,0 @@
-'use server';
-
-import { TaskDTO, TaskRepository } from "@/model/Task";
-
-export async function addTodoAction(newTodo: TaskDTO) {
-    const columnRepository = new TaskRepository();
-    columnRepository.create(newTodo);
-  }
