--- src/templates/actions/editTodoAction.ts
@@ -1,8 +1,0 @@
-'use server';
-
-import { TaskDTO, TaskRepository } from "@/model/Task";
-
-export async function editTodoAction(editedTodo: TaskDTO) {
-    const columnRepository = new TaskRepository();
-    columnRepository.update(editedTodo);
-  }
